Splitting data and master nodes

+++ templates/user_data.sh
@@ -8,10 +8,6 @@ cat <<'EOF' >>/etc/elasticsearch/elasticsearch.yml
 
 cluster.name: ${es_cluster}
 
-path:
-  logs: ${elasticsearch_logs_dir}
-  data: ${elasticsearch_data_dir}
-
 discovery.zen.minimum_master_nodes: ${minimum_master_nodes}
 network.host: _ec2:privateIpv4_
 
@@ -24,6 +20,15 @@ discovery:
     ec2.host_type: private_ip
     ec2.tag.Cluster: ${es_environment}
     ec2.availability_zones: ${availability_zones}
+
+# only data nodes should have ingest and http capabilities
+node.master: ${master}
+node.data: ${data}
+node.ingest: ${data}
+http.enabled: ${data}
+
+path.logs: ${elasticsearch_logs_dir}
+
 EOF
 
 cat <<'EOF' >>/etc/security/limits.conf
@@ -41,13 +46,18 @@ sudo sed -i "s/^-Xmx2g/-Xmx${heap_size}/" /etc/elasticsearch/jvm.options
 sudo sed -i "s/#LimitMEMLOCK=infinity/LimitMEMLOCK=infinity/" /usr/lib/systemd/system/elasticsearch.service
 
 # Storage
-sudo mkfs -t ext4 ${volume_name}
-sudo mkdir -p ${elasticsearch_data_dir}
 sudo mkdir -p ${elasticsearch_logs_dir}
-sudo mount ${volume_name} ${elasticsearch_data_dir}
-sudo echo "${volume_name} ${elasticsearch_data_dir} ext4 defaults,nofail 0 2" >> /etc/fstab
-sudo chown -R elasticsearch:elasticsearch ${elasticsearch_data_dir}
 sudo chown -R elasticsearch:elasticsearch ${elasticsearch_logs_dir}
+
+# we are assuming volume is declared and attached when data_dir is passed to the script
+if [ -n "${elasticsearch_data_dir}" ]; then
+    sudo sed -i '$ a path.data: ${elasticsearch_data_dir}' /etc/elasticsearch/elasticsearch.yml
+    sudo mkfs -t ext4 ${volume_name}
+    sudo mkdir -p ${elasticsearch_data_dir}
+    sudo mount ${volume_name} ${elasticsearch_data_dir}
+    sudo echo "${volume_name} ${elasticsearch_data_dir} ext4 defaults,nofail 0 2" >> /etc/fstab
+    sudo chown -R elasticsearch:elasticsearch ${elasticsearch_data_dir}
+fi
 
 # Start Elasticsearch
 sudo /bin/systemctl daemon-reload
